Moved license functions to license.py
